Submit includes jire isssue (SENSOR-35)
Also includes style changes to the Data.aspx pages
As well as substantial amounts of javascript code commenting - after we establish its safe to remove it will be deleted.

+++ TagCloud/Service References/TagService/Reference.cs
@@ -24,7 +24,7 @@ namespace TagCloud.TagService {
         
         private string NameField;
         
-        private int WeightField;
+        private double WeightField;
         
         [System.Runtime.Serialization.DataMemberAttribute()]
         public bool IsSelected {
@@ -53,7 +53,8 @@ namespace TagCloud.TagService {
         }
         
         [System.Runtime.Serialization.DataMemberAttribute()]
-        public int Weight {
+        public double Weight
+        {
             get {
                 return this.WeightField;
             }
